Crate anchor_attribute_account

Macros

declare_id

Defines the program’s ID. This should be used at the root of all Anchor based programs.

pubkey

Convenience macro to define a static public key.

Attribute Macros

account

An attribute for a data structure representing a Solana account.

zero_copy

A data structure that can be used as an internal field for a zero copy deserialized account, i.e., a struct marked with #[account(zero_copy)].

Derive Macros

ZeroCopyAccessor